requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
Django==3.0.3
django-telegrambot==1.0.1
future==0.18.2
idna==2.9
psycopg2-binary==2.8.4
pycparser==2.19
python-telegram-bot==12.4.2
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.8
